Extracted exception handling delegate to separate static class
