+++ backend/BrimSchedule.API/Startup.cs
@@ -4,12 +4,9 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Security.Claims;
-using System.Text.Json;
 using System.Threading.Tasks;
-using BrimSchedule.API.Models;
 using BrimSchedule.API.Services;
 using BrimSchedule.API.SwaggerConfiguration;
-using BrimSchedule.Application.Logging;
 using BrimSchedule.Domain.Models;
 using BrimSchedule.Persistence.EF;
 using FirebaseAdmin;
@@ -17,9 +14,7 @@ using FirebaseAdmin.Auth;
 using Google.Apis.Auth.OAuth2;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
 using Microsoft.Extensions.Configuration;
@@ -174,32 +169,8 @@ namespace BrimSchedule.API
 			{
 				app.UseExceptionHandler(errorApp =>
 				{
-					errorApp.Run(async context =>
-					{
-						context.Response.StatusCode = 500;
-						context.Response.ContentType = "application/json";
-
-						var exceptionHandlerPathFeature = context.Features.Get<IExceptionHandlerPathFeature>();
-						var exception = exceptionHandlerPathFeature.Error;
-
-						const string defaultErrorMessage = "Server error occured";
-
-						var errorMessage = exception is UserFriendlyException ? exception.Message : defaultErrorMessage;
-						var errorId = Guid.NewGuid();
-						var serviceError = new ServiceError
-						{
-							ErrorId = errorId,
-							ErrorMessage = errorMessage,
-						};
-
-						var serviceErrorJson = JsonSerializer.Serialize(serviceError);
-						await context.Response.WriteAsync(serviceErrorJson);
-
-						var logger = context.RequestServices.GetService<ILoggingManager>();
-						logger?.Error(serviceErrorJson, exception);
-					});
+					errorApp.Run(async context => { await ExceptionHandler.HandleGlobalExceptions(context); });
 				});
-				app.UseHsts();
 			}
 
 			app.UseRouting();
@@ -224,6 +195,7 @@ namespace BrimSchedule.API
 			{
 				app.UseCors();
 				app.UseHttpsRedirection();
+				app.UseHsts();
 			}
 
 			var firebaseConfigFilePath =
